# Performance improvements to conversation list filtering